complete edit wishlist functionality on frontend
