+++ src/components/WishlistForm/WishlistForm.jsx
@@ -1,18 +1,32 @@
 import styles from "./WishlistForm.module.css"
+import { useState } from "react"
+import { useLocation } from "react-router-dom"
 
 const WishlistForm = (props) => {  
+  const { state } = useLocation()
+  const [wishlistForm, setWishlistForm] = useState(state)
+
+  const handleChange = ({ target }) => {
+    setWishlistForm({ ...wishlistForm, [target.name]: target.value })
+  }
+
+  const handleSubmit = (e) => {
+    e.preventDefault()
+		props.handleUpdateWishlist(wishlistForm)
+  }
+
   return ( 
     <>
-      <form onSubmit={props.handleSubmit} className={styles.itemForm}>
+      <form onSubmit={handleSubmit} className={styles.itemForm}>
         <label htmlFor="name-input">Name</label>
         <input
           required
           type="text"
           name="name"
           id="name-input"
-          value={props.wishlistForm.name}
+          value={wishlistForm.name}
           placeholder="Name"
-          onChange={props.handleChange}
+          onChange={handleChange}
         />
         <label htmlFor="description-input">Description</label>
 				<textarea
@@ -20,9 +34,9 @@ const WishlistForm = (props) => {
           type="text"
           name="description"
           id="description-input"
-          value={props.wishlistForm.description}
+          value={wishlistForm.description}
           placeholder="Description"
-          onChange={props.handleChange}
+          onChange={handleChange}
         />
         <button type="submit">SUBMIT</button>
       </form>
